norminate & finish everything except ft_lst

+++ src/ft_string/ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: kyung-ki <kyung-ki@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/20 12:30:06 by kyung-ki          #+#    #+#             */
-/*   Updated: 2023/10/20 19:20:05 by kyung-ki         ###   ########.fr       */
+/*   Updated: 2023/10/21 21:24:27 by kyung-ki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,10 +16,10 @@
 char	*ft_strdup(const char *s1)
 {
 	char	*tmp;
-	int		i;
+	size_t	i;
 
 	tmp = (char *)malloc(ft_strlen(s1) + 1);
-	if(!tmp)
+	if (!tmp)
 		return (0);
 	i = -1;
 	while (++i < ft_strlen(s1))
